requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
pydantic
pydantic-settings
python-dotenv
redis
langgraph
langchain
langchain-openai
httpx
